incorrect type location caused utils package error
