+++ vuu-ui/showcase/src/examples/VuuFeatures/FilterTableFeature.examples.tsx
@@ -1,11 +1,15 @@
 import { getSchema } from "@finos/vuu-data-test";
 import { FlexboxLayout, LayoutProvider, View } from "@finos/vuu-layout";
-import { Feature, FeatureProps, useWorkspace } from "@finos/vuu-shell";
+import { Feature, useWorkspace } from "@finos/vuu-shell";
 import { useCallback, useState } from "react";
 import { FilterTableFeature } from "../../features/FilterTable.feature";
 import { VuuBlotterHeader } from "./VuuBlotterHeader";
 import { JsonTable } from "@finos/vuu-datatable";
-import { type JsonData, registerComponent } from "@finos/vuu-utils";
+import {
+  type FeatureProps,
+  type JsonData,
+  registerComponent,
+} from "@finos/vuu-utils";
 
 registerComponent("FilterTableFeature", FilterTableFeature, "view");
